move build_mask to utils
