+++ experiments/weaklensing/main.py
@@ -1,11 +1,8 @@
 import numpy as np
 import argparse
-from os import path
 import datetime
-import random
 import pyssht
 from pys2let import lm_hp2lm
-from astropy.coordinates import SkyCoord
 import healpy as hp
 
 from pxmcmc.measurements import WeakLensing
@@ -14,39 +11,7 @@ from pxmcmc.forward import ForwardOperator
 from pxmcmc.mcmc import PxMCMCParams, MYULA, PxMALA, SKROCK
 from pxmcmc.prior import S2_Wavelets_L1
 from pxmcmc.saving import save_mcmc
-
-
-def build_mask(L):
-    """"
-    Builds a mask for the galactic plane and ecliptic
-    0 at positions to be masked
-    i.e. to apply mask do map * mask
-
-    Mask in MW format
-    """
-    mask = np.ones(pyssht.sample_shape(L))
-    thetas, phis = pyssht.sample_positions(L)
-    for i, t in enumerate(thetas):
-        for j, p in enumerate(phis):
-            if np.abs(90 - np.degrees(t)) < 20:
-                mask[i, j] = 0
-
-    thetaarray, phiarray = pyssht.sample_positions(L, Grid=True)
-    thetaarray = np.degrees(thetaarray) - 90
-    phiarray = np.degrees(phiarray) - 180
-
-    c = SkyCoord(phiarray, thetaarray, unit="deg")
-    d = c.transform_to("galactic")
-    degm = np.abs(d.b.degree)
-    for i in range(L):
-        for j in range(2 * L - 1):
-            rand = random.random()
-            if rand < 0.1:
-                mask[i, j] = 0
-            if degm[i, j] < 20:
-                mask[i, j] = 0
-
-    return mask
+from pxmcmc.utils import build_mask
 
 
 def load_gammas(kappa_fits_file: str, L: int, wl: WeakLensing):
